Simplify the registration flow
Put the burden of knowing when registration is needed on Handy. Instead of a new member first sending in their name, then their email address, and finally an event token, recognize when a new member is attempting to check in with an event token and ask for their name and email address.