Fix sporadic test failures
Because the implementation of Dictionary.words involves memoizing the list of dictionary words in
share/dictionary.txt, and tests are run in random order, EventsController#create could occasionally
be tested before Dictionary. When this happened, the words in share/dictionary.txt were loaded and
memoized, making the stubs for File.readlines in DictionaryTest ineffective.